refactor Storage
